% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_time_periods.R
\name{assign_time_periods}
\alias{assign_time_periods}
\title{Add a date_factor column to input data.frame that groups consecutive dates.}
\usage{
assign_time_periods(tseries_tbl, time_period)
}
\arguments{
\item{time_period}{}
}
\description{
To be implemented. For now, the code expects a data frame that already has a date_factor column.
}
