% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_causal_estimates.R
\name{breakup_est_by_treatment_group}
\alias{breakup_est_by_treatment_group}
\title{Aggregates estimates for (time-period, treatment) pairs over treatment groups. Useful when treatments can be classified in a fixed number of groups.}
\usage{
breakup_est_by_treatment_group(ctr_df, estimate_colname)
}
\arguments{
\item{estimate_colname}{}
}
\value{
A data.frame containing mean causal estimates for each treatment group.
}
\description{
Aggregates estimates for (time-period, treatment) pairs over treatment groups. Useful when treatments can be classified in a fixed number of groups.
}
