% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_checks.R
\name{check_distribution_splitdoors_by_group}
\alias{check_distribution_splitdoors_by_group}
\title{Compare distribution of all treatments and treatments with a valid split-door estimate.}
\usage{
check_distribution_splitdoors_by_group(causal_estimate_df,
  ordered_treatment_groups_vec = NULL, do_plot = TRUE)
}
\arguments{
\item{do_plot}{}
}
\description{
Compare distribution of all treatments and treatments with a valid split-door estimate.
}
