% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_checks.R
\name{check_robustness_independence_threshold}
\alias{check_robustness_independence_threshold}
\title{Check robustness to independence threshold.}
\usage{
check_robustness_independence_threshold(tseries_df,
  independence_thresholds = c(0.001, 0.005, 0.01, 0.05, 0.1, 0.15, 0.2),
  do_plot = TRUE, ...)
}
\arguments{
\item{..}{}
}
\value{
A data.frame containing average causal estimates with different user-provided values of the independence threshold.
}
\description{
Check the variation of causal estimate as the 
threshold for independence used by split-door criterion is varied.
}
