% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdoor.R
\name{compare_splitdoor_correlational_estimate}
\alias{compare_splitdoor_correlational_estimate}
\title{Compare a baseline correlational measure to causal estimates using the split-door criterion.}
\usage{
compare_splitdoor_correlational_estimate(splitdoor_estimates_df,
  correlational_estimates_df, by_group = TRUE,
  estimate_colname = "causal_estimate")
}
\arguments{
\item{splitdoor_estimates_df}{}

\item{correlational_estimates_df}{}

\item{by_group}{}

\item{estimate_colname}{}
}
\value{
If by_group is TRUE, a data.frame containing the mean estimates for each group using split-door and the baseline method.
}
\description{
Compare a baseline correlational measure to causal estimates using the split-door criterion.
}
