% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_causal_effect.R
\name{compute_pairwise_causal_effect}
\alias{compute_pairwise_causal_effect}
\title{Helper function. Aggregates timeseries values over selected time-periods and computes the causal effect as a ratio.}
\usage{
compute_pairwise_causal_effect(valid_tseries_tbl, is_valid_splitdoor = TRUE)
}
\arguments{
\item{is_valid_splitdoor}{}
}
\value{
A data.frame containing the causal estimate for each (treatment, outcome) pair.
Causal estimate is NA if (treatment, outcome, time-period) triplet does not satisfy the split-door criterion.
}
\description{
Helper function. Aggregates timeseries values over selected time-periods and computes the causal effect as a ratio.
}
