% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdoor.R
\name{correlational_estimate}
\alias{correlational_estimate}
\title{Title}
\usage{
correlational_estimate(tseries_df, test_time_period = NULL,
  fn_add_time_period_factor = assign_time_periods)
}
\arguments{
\item{fn_add_time_period_factor}{}
}
\description{
Title
}
