% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_test.R
\name{dcor_independence_test}
\alias{dcor_independence_test}
\title{Randomization Test for independence between pairs of time-series.}
\usage{
dcor_independence_test(tseries_tbl, p_value)
}
\arguments{
\item{p_value}{}
}
\value{
A data.frame containing test results for each (treatment, outcome) pair.
Probability of independence of the two time-series is provided along with Pass/Fail result of the test based on the user-provided independence_threshold.
}
\description{
Uses randomization test based on distance correlation.
}
