% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_causal_effect.R
\name{est_causal_effect}
\alias{est_causal_effect}
\title{Compute effect of treatment on outcome over selected subsets of data, given knowledge about which (treatment, outcome) pairs satisfy the split-door criterion.}
\usage{
est_causal_effect(tseries_tbl_with_indep_info, method = "simple")
}
\arguments{
\item{method}{}
}
\value{
A data.frame containing causal estimates for each (treatment, outcome) pair.
}
\description{
Compute effect of treatment on outcome over selected subsets of data, given knowledge about which (treatment, outcome) pairs satisfy the split-door criterion.
}
