% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_causal_effect.R
\name{est_naive}
\alias{est_naive}
\title{A baseline correlational measure of the effect of treatment on outcome.}
\usage{
est_naive(tseries_tbl)
}
\arguments{
\item{tseries_tbl}{}
}
\value{
A data.frame containing estimates for each (treatment, outcome) pair.
}
\description{
Computes the ratio of the outcome and treatment time-series.
}
