% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_test.R
\name{fisher_independence_test}
\alias{fisher_independence_test}
\title{Fisher's test for independence between pairs of discrete time-series.}
\usage{
fisher_independence_test(tseries_tbl, p_value, num_discrete_levels)
}
\arguments{
\item{num_discrete_levels}{}
}
\value{
A data.frame containing test results for each (treatment, outcome) pair.
Probability of independence of the two time-series is provided along with Pass/Fail result of the test based on the user-provided independence_threshold.
}
\description{
Uses fisher's test for discrete variables.
}
