% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_checks.R
\name{inspect_splitdoor_pairs}
\alias{inspect_splitdoor_pairs}
\title{Inspect manually timeseries for  <treatment, aux_outcome> pairs that are returned by the split-door criterion.}
\usage{
inspect_splitdoor_pairs(treatment_outcome_pairs, tseries_df,
  n_treatment_outcome_pairs = NULL)
}
\arguments{
\item{n_treatment_outcome_pairs}{}
}
\description{
Inspect manually timeseries for  <treatment, aux_outcome> pairs that are returned by the split-door criterion.
}
