% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_ctr_compare}
\alias{plot_ctr_compare}
\title{Compare correlational and split-door estimates for different treatment groups.}
\usage{
plot_ctr_compare(compare_estimates, min_group_frequency,
  filter_out_zeroyd = TRUE, yname = "Click-through Estimate",
  ylim_max = 0.2)
}
\arguments{
\item{ylim_max}{}
}
\description{
Compare correlational and split-door estimates for different treatment groups.
}
