% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_estimate_by_indep_threshold}
\alias{plot_estimate_by_indep_threshold}
\title{Plot mean split-door estimate for different values of the independence threshold.}
\usage{
plot_estimate_by_indep_threshold(mean_est_df, saveto_filename = NULL)
}
\arguments{
\item{saveto_filename}{}
}
\description{
Plot mean split-door estimate for different values of the independence threshold.
}
