% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_num_splitdoors_by_indep_threshold}
\alias{plot_num_splitdoors_by_indep_threshold}
\title{Plot number of unique treatments that have at least one time-period when the split-door criterion is satisfied.}
\usage{
plot_num_splitdoors_by_indep_threshold(mean_est_df, saveto_filename = NULL)
}
\arguments{
\item{saveto_filename}{}
}
\value{

}
\description{
Compare this value as the independence threshold varies.
}
