% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_pgroup_distribution}
\alias{plot_pgroup_distribution}
\title{Plot distribution of treatments w.r.t. pre-specified treatment groups that have at least one valid time-period when the split-door criterion is satisfied.}
\usage{
plot_pgroup_distribution(pgroup_plotdata)
}
\arguments{
\item{pgroup_plotdata}{}
}
\description{
Also compares to overall distribution of treatments over treatment groups.
}
