% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_treatment_aux_outcome_timeseries}
\alias{plot_treatment_aux_outcome_timeseries}
\title{Plot examples of <treatment,outcome> time-series.}
\usage{
plot_treatment_aux_outcome_timeseries(plot_df)
}
\arguments{
\item{plot_df}{}
}
\description{
Used to manually inspect timeseries for <treatment,outcome> pairs returned by
the split-door criterion.
}
