% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_test.R
\name{gen_cor_samples}
\alias{gen_cor_samples}
\title{Randomization Test for independence between pairs of time-series.}
\usage{
gen_cor_samples(x_vec, y_vec, cor_metric)
}
\arguments{
\item{p_value}{}
}
\value{
A data.frame containing test results for each (treatment, outcome) pair.
Probability of independence of the two time-series is provided along with Pass/Fail result of the test based on the user-provided independence_threshold.
}
\description{
Uses randomization test based on distance correlation.
}
